[2/2] Enable torch-xpu-ops on Native Windows (take 2) (#458)
As part of https://github.com/pytorch/pytorch/issues/126719, we are
enabling torch-xpu-ops on Native Windows.

The enabling is separated into two parts:

1. CMake changes to build on Windows
https://github.com/intel/torch-xpu-ops/pull/287
2. Op changes to build on Windows
https://github.com/intel/torch-xpu-ops/pull/458

This PR is the second part of the above. This PR:

- Fixes Windows or MSVC issues on some existing operator
implementations.

Currently, there's no CI/CD for Windows to gatekeep PRs that break on
Windows. Hence this PR manually rebases to main, and fixes Window/MSVC
issues from merged PRs. Once CI/CD is added for Windows, PRs will be
gatekept to catch issues before merge.

Co-authored-by: Feng Yuan <feng1.yuan@intel.com>