Export EmptyTensor.h to public (#935)
# Motivation
Export `<ATen/xpu/EmptyTensor.h>` to public (the counterpart location of
CUDA is `${AT_INSTALL_INCLUDE_DIR}/ATen/cuda`） as it is used in codegen,
refer to
https://github.com/pytorch/pytorch/pull/130082/files#r1696016201
Otherwise, IPEX finds it hard to reuse this codegen due to the non-exist
file of `<ATen/xpu/EmptyTensor.h>`.
The same reason applies to libtorch, which is one pure C++ code.