Fix batch norm vectorize path accuracy issue by enforcing shape alignment (#1238)
When to implement a kernel for all shapes with vectorized LD/ST, we have
to handle non-aligned head (base address) and short tail (tail < vector
size). Before the commit, there was lack of head handling which leaded
to non-vector-size-aligned vectorized ST.
The fixing enforces shapes for vectorize path. The vectorized kernel
could be called only in the case in which feature dim could be divided
by vector size. There will always be aligned head and exact tail.

---------

Co-authored-by: mengfei25 <mengfei.li@Intel.com>