Building System: Enable device code compression (Lin/Win) (#1148)
This is the first stage to align binary distribution with CUDA. That is
only `libtorch_cuda.so` at all. The first stage helps on combining all
kernel libraries into one. After the commit, we will have two libraries
`libtorch_xpu.so` and `libtorch_xpu_ops_sycl_kernel.so`, and the commit
also helps us make the bin size on par as CUDA. In the next stage, we
will
1. Combine all XPU libraries into one, `libtorch_xpu.so`.
2. Revert lin/win divergence.

---------

Signed-off-by: Feng Yuan <feng1.yuan@intel.com>
Co-authored-by: chunhuanMeng <chunhuan.meng@intel.com>
Co-authored-by: chunhuanMeng <105194461+chunhuanMeng@users.noreply.github.com>