Align latest dist.barrier test change (#1639)
Refer https://github.com/pytorch/pytorch/pull/149144, Currently,
`dist.barrier` accepts `device_ids` as a parameter that doesn't have to
be a list. When `device_ids` is not provided or another value is passed,
`barrier` will use the device associated with the process group at
initialization to perform the synchronization.