Enhance build flags for Debug and RelWithDebInfo configurations (#1671)
Updates the build configuration in `cmake/BuildFlags.cmake` for
different build types. The changes focus on improving debugging
capabilities and adding specific flags for `Debug` and `RelWithDebInfo`
build types.

### Debugging and Optimization Flags:


* Added a condition in `CMakeLists.txt` to automatically enable
`BUILD_SEPARATE_OPS` when the build type is `Debug` or `RelWithDebInfo`.
* Consolidated `SYCL_HOST_FLAGS` for the `Debug` build type to include
`-g`, `-fno-omit-frame-pointer`, and `-O0`, and added a new
configuration for the `RelWithDebInfo` build type with `-g` and `-O2`.
* Updated `SYCL_KERNEL_OPTIONS` to add `-g -O0
-Rno-debug-disables-optimization` for `Debug` builds and
`-gline-tables-only -O2` for `RelWithDebInfo` builds.

---------

Co-authored-by: Yutao Xu <yutao.xu@intel.com>