Add an option for NAN check for xccl (#1756)
Refer from https://github.com/pytorch/pytorch/pull/125726,
https://github.com/pytorch/pytorch/pull/135414.
Add nan check for xccl.
why we need to stop communication from spreading NaNs?
"technically if we can be sure which rank (or, even which host) detected
the first nan, then its OK to let the nan spread to some other hosts.
but in practice i dont know if we have good enough way to align our logs
on different hosts, so if we let the nan spread to a few other hosts we
may lose track of which one was first”

---------

Co-authored-by: mengfei25 <mengfei.li@Intel.com>