build: enable SYCL warnings on Linux (#2096)
Some versions of DPC++ compiler pass paths to SYCL headers as user
include paths (`-I`) rather than system paths (`-isystem`). This makes
host compiler to report warnings encountered in the SYCL headers, such
as deprecated warnings, even if warned API is not actually used in the
program. We expect that this issue will be addressed in the later
version of DPC++ compiler. To workaround the issue we wrap paths to SYCL
headers in `-isystem`.

disable_ut
disable_e2e
disable_distributed

CC: @EikanWang @chuanqi129

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>