build: drop --force-rpath from patchelf to favor runpath
Fixes: #2349

`RPATH` is handled differently from `RUNPATH` in a way that
`LD_LIOBRARY_PATH` is searched before `RPATH` and after `RUNPATH`.
With the oneAPI we step into backward incompatibility of Unified
Runtime component (`libur_loader.so`) which does not have stable API
and ABI. This results in the loading issue when user sources later
oneAPI environment (which is supposed to be backward compatible).
However, due to usage of `RPATH` the `libsycl.so` is picked from
the wheel packages, but `libur_loader.so` is picked from the sourced
oneAPI distribution.

```
ImportError: /opt/intel/oneapi/2025.2/lib/libur_loader.so.0: version `LIBUR_LOADER_0.11' not found (required by /home/dvrogozh/pt2.8/lib/python3.12/site-packages/torch/lib/../../../../libsycl.so.8)
```

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>